requirements.txt
python-telegram-bot==20.7
flask==3.0.0
gspread==6.0.2
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.1
google-auth==2.26.2